* Apply http://opensimulator.org/mantis/view.php?id=3151 * Fixes NHibernate overflow exception when saving some objects (under at least PostgreSQL 8.3) * Thanks Tommil!

+++ OpenSim/Data/NHibernate/Resources/MsSql2005Dialect/001_InventoryStore.sql
@@ -1,7 +1,7 @@
 create table InventoryFolders (
   ID NVARCHAR(255) not null,
    Type SMALLINT null,
-   Version SMALLINT null,
+   Version INT null,
    ParentID NVARCHAR(255) null,
    Owner NVARCHAR(255) null,
    Name NVARCHAR(64) null,
